holy shit pico-arduino-compat is NOT worth it!!!!!
what a MESS! i give up trying to make it build in a sane way.

+++ build_and_flash.ps1
@@ -9,6 +9,10 @@ function ThrowOnNativeFailure {
     }
 }
 
+if (!(Test-Path $UploadPath)) {
+    throw "UploadPath $UploadPath does not exist"
+}
+
 bazel build -c opt //tplp:firmware.uf2
 ThrowOnNativeFailure
 Copy-Item bazel-bin/tplp/firmware.uf2 $UploadPath
